+++ src/extension.ts
@@ -26,8 +26,8 @@ export function activate(context: vscode.ExtensionContext): void {
     context.subscriptions.push(vscode.commands.registerCommand('jetty.server.stop.context', (server: JettyServer) => { jettyServerController.stopServer(server); }));
     context.subscriptions.push(vscode.commands.registerCommand('jetty.server.delete', (server: JettyServer) => { jettyServerController.deleteServer(server); }));
     context.subscriptions.push(vscode.commands.registerCommand('jetty.server.delete.context', (server: JettyServer) => { jettyServerController.deleteServer(server); }));
-    context.subscriptions.push(vscode.commands.registerCommand('jetty.war.deploy', (uri: vscode.Uri) => { jettyServerController.deployWarPackage(uri); }));
-    context.subscriptions.push(vscode.commands.registerCommand('jetty.war.debug', (uri: vscode.Uri) => { jettyServerController.debugWarPackage(uri); }));
+    context.subscriptions.push(vscode.commands.registerCommand('jetty.war.run', (uri: vscode.Uri) => { jettyServerController.runWarPackage(uri); }));
+    context.subscriptions.push(vscode.commands.registerCommand('jetty.war.debug', (uri: vscode.Uri) => { jettyServerController.runWarPackage(uri); }));
     context.subscriptions.push(vscode.commands.registerCommand('jetty.tree.refresh', (server: JettyServer) => { jettyServerTree.refresh(server); }));
 }